refac: move non lsp code into separate folder
