+++ lib/common/format.js
@@ -14,7 +14,7 @@ export default function format(input: string): string {
       // if some valid javascript
       // example: type Props = { value: string };
       return runPrettier(input);
-    } catch (err) {
+    } catch {
       // cases which dont need formatting or we cant run prettier
       // example: `class XYZ`, `import name`
       return input;
@@ -29,6 +29,6 @@ function runPrettier(code: string): string {
     // vscode uses max-width 500px on hover ui
     // prettier better handles wrapping
     // @TODO: use current fontSize to decide value (??)
-    printWidth: 60
+    printWidth: 60,
   });
 }
