+++ lib/common/useLSP.js
@@ -1,0 +1,6 @@
+/* @flow */
+import * as vscode from 'vscode';
+
+export default function useLSP() {
+  return vscode.workspace.getConfiguration('flow').get('useLSP');
+}
