+++ lib/flowNonLSP/flowCompletion.js
