+++ lib/flowNonLSP/flowDeclaration.js
