+++ lib/flowNonLSP/flowDiagnostics.js
