+++ lib/flowNonLSP/flowLogging.js
@@ -2,9 +2,11 @@
  * Adds a gloabl that is used inside consoleAppender.js to output console messages
  * to the user, instead of to the developer console.
  */
+import * as vscode from 'vscode';
+
 export function setupLogging(context): void {
-  const channel = vscode.window.createOutputChannel("Flow");
-  vscode.commands.registerCommand("flow.show-output", () => channel.show());
+  const channel = vscode.window.createOutputChannel('Flow');
+  vscode.commands.registerCommand('flow.show-output', () => channel.show());
   global.flowOutputChannel = channel;
   context.subscriptions.push(channel);
 }
