+++ lib/flowNonLSP/pkg/commons-node/ScribeProcess.js
