+++ lib/flowNonLSP/pkg/commons-node/singleton.js
