+++ lib/flowNonLSP/pkg/commons-node/stream.js
