+++ lib/flowNonLSP/pkg/nuclide-logging/lib/consoleAppender.js
