+++ lib/flowNonLSP/pkg/nuclide-remote-uri/package.json
