+++ lib/flowNonLSP/utils/index.js
@@ -4,10 +4,9 @@ export {
   checkNode,
   getFlowVersion,
   isFlowEnabled,
-  useLSP,
   isRunOnEditEnabled,
   shouldStopFlowOnExit,
   isFlowStatusEnabled
 } from './util';
 
-export { default as format } from './format';
+export { default as format } from '../../common/format';
