+++ lib/flowNonLSP/utils/util.js
@@ -19,10 +19,6 @@ export function isFlowEnabled() {
 	return workspace.getConfiguration('flow').get('enabled')
 }
 
-export function useLSP() {
-	return workspace.getConfiguration('flow').get('useLSP')
-}
-
 export function isFlowStatusEnabled():boolean {
 	return workspace.getConfiguration('flow').get('showStatus')
 }
