+++ lib/index.js
@@ -1,11 +1,10 @@
 /* @flow */
 import { ExtensionContext } from 'vscode';
-import { useLSP } from './utils';
+import useLSP from './common/useLSP';
 
 export function activate(context: ExtensionContext) {
   if (useLSP()) {
     return require('./flowLSP').activate(context);
-  } else {
-    require('./flowMain').activate(context);
   }
+  return require('./flowNonLSP').activate(context);
 }
